(refactor) change tests and helpers to work with new models destination
